move region- and feature-extractor to separate files
